update require after move
