+++ core/utils.js
@@ -5,8 +5,11 @@ const logger = winston.loggers.get('gomngr');
 var CONFIG = require('config');
 var GENERAL_CONFIG = CONFIG.general;
 const MAILER = CONFIG.general.mailer;
-var notif = require('../routes/notif_'+MAILER+'.js');
 const marked = require('marked');
+
+const notif = require('../core/notif_'+MAILER+'.js');
+
+/* TODO : SHOULD SPLIT THIS FILE, one for mongo, one for redis, on for mail ..... */
 
 const MongoClient = require('mongodb').MongoClient;
 var mongodb = null;
