+++ routes/logs.js
@@ -1,10 +1,10 @@
-var express = require('express');
+ar express = require('express');
 var router = express.Router();
 var fs = require('fs');
 var CONFIG = require('config');
 var GENERAL_CONFIG = CONFIG.general;
 
-var utils = require('./utils');
+const utils = require('../core/utils.js');
 
 router.get('/log', async function(req, res){
     if(! req.locals.logInfo.is_logged) {
